feat: add IoC and restructure code
