+++ src/Daf.Oche.Console/Daf.Oche.Cli.csproj
@@ -8,11 +8,12 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\Daf.Oche.Domain\Daf.Oche.Domain.csproj" />
+    <Compile Include="Oche.cs" />
+    <Compile Include="Program.cs" />
   </ItemGroup>
 
   <ItemGroup>
-    <PackageReference Include="Microsoft.Extensions.DependencyInjection" Version="9.0.0" />
+    <ProjectReference Include="..\Daf.Oche.Domain\Daf.Oche.Domain.csproj" />
   </ItemGroup>
 
 </Project>
