+++ src/Daf.Oche.Domain/Daf.Oche.Domain.csproj
@@ -10,4 +10,7 @@
 			<_Parameter1>$(MSBuildProjectName).Tests</_Parameter1>
 		</AssemblyAttribute>
 	</ItemGroup>
+	<ItemGroup>
+	  <PackageReference Include="Microsoft.Extensions.DependencyInjection" Version="9.0.0" />
+	</ItemGroup>
 </Project>
